Added a simple os distribution detection
As we can't rely on the facts, because the OS might be changed
during an install, we need to figure out the new OS after the
installation. This is limited to Debian and Red Hat based
distributions. We tested Arch and it had python installed, SuSE
is currently untested, but should be detected.