Added interpreter workaround for ansible <2.8
Since ansible 2.8 the python interpreter can be detected
automatically. But with ansible versions less than 2.8, this
feature isn't available. Since the rescue system is based on
Debian, the target or "old" os might be CentOS or such, they might
have different interpreters. With this fix the interpreter for the
rescue image is set by default to /usr/bin/python while for all
non rescue system calls, the interpreter set in the inventory or
automatically found will be used.